Moving elevations to new design-token package structure
Deprecate shadows
Structure elevations a bit
