+++ packages/design-tokens/src/index.ts
@@ -32,6 +32,7 @@ import {
 export const shouldForwardProp = createShouldForwardProp([...props])
 
 export * from './color'
+export * from './elevation'
 export type {
   LegacySpaceRamp,
   SpaceRamp,
