+++ packages/design-tokens/src/system/index.ts
@@ -91,7 +91,6 @@ export type { DensityProp, DensityRamp } from './density'
 export { cursor } from './cursor'
 export type { CursorProps } from './cursor'
 export { Easings } from './easings'
-export { Elevations } from './elevations'
 export { RadiusSizes, Radii } from './radii'
 export { Shadows } from './shadows'
 export * from './transitions'
