+++ packages/design-tokens/src/system/shadows.ts
@@ -26,4 +26,7 @@
 
 export type ElevationLayers = '1' | '2' | '3' | '4' | '5' | '6'
 
+/**
+ * @deprecated - Use `Elevations` instead
+ */
 export type Shadows = { [K in ElevationLayers]: string }
