+++ packages/design-tokens/src/tokens/shadows.ts
@@ -26,6 +26,9 @@
 
 import { Shadows } from '../system'
 
+/**
+ * @deprecated - Use `elevations` instead
+ */
 export const shadows: Shadows = {
   1: '0px 1px 8px rgba(0, 0, 0, 0.08), 0px 1px 1px rgba(0, 0, 0, 0.05)',
   2: '0 2px 12px rgba(0, 0, 0, 0.11), 0 1px 4px rgba(0, 0, 0, 0.04)',
